[backport] Move ShHeap::used increment out of locked allocation path

+++ src/share/vm/gc_implementation/shenandoah/shenandoahHeap.hpp
@@ -208,7 +208,7 @@ private:
 
   ShenandoahWorkGang* _workers;
 
-  volatile size_t _used;
+  volatile jlong _used;
   volatile size_t _committed;
 
   MarkBitMap _verification_bit_map;
